chore(node): move utxo list ser/deser functions
