+++ demo/midgard-node/src/database/utils/common.ts
@@ -2,6 +2,7 @@ import { Data, Effect } from "effect";
 import { Database } from "@/services/database.js";
 import { SqlClient, SqlError } from "@effect/sql";
 import * as SDK from "@al-ft/midgard-sdk";
+import { CML, coreToUtxo, UTxO, utxoToCore } from "@lucid-evolution/lucid";
 
 export const retrieveNumberOfEntries = (
   tableName: string,
@@ -79,3 +80,84 @@ export const sqlErrorToDatabaseError = (
             cause: error,
           }),
   );
+
+/**
+ * Serializes a UTxO list by converting each UTxO to core and storing its CBOR
+ * bytes as hex inside a JSON array.
+ */
+export const serializeUTxOsForStorage = (
+  utxos: readonly UTxO[],
+): Effect.Effect<Buffer, SDK.CborSerializationError | SDK.CmlUnexpectedError> =>
+  Effect.gen(function* () {
+    const serializedEach = yield* Effect.forEach(
+      utxos,
+      (utxo) =>
+        Effect.try({
+          try: () =>
+            Buffer.from(utxoToCore(utxo).to_cbor_bytes()).toString("hex"),
+          catch: (e) =>
+            new SDK.CmlUnexpectedError({
+              message: `Failed to serialize UTxO to core CBOR`,
+              cause: e,
+            }),
+        }),
+      { concurrency: "unbounded" },
+    );
+    return yield* Effect.try({
+      try: () => Buffer.from(JSON.stringify(serializedEach), "utf8"),
+      catch: (e) =>
+        new SDK.CborSerializationError({
+          message: `Failed to serialize UTxO list payload`,
+          cause: e,
+        }),
+    });
+  });
+
+/**
+ * Deserializes UTxO list payload produced by `serializeUTxOsForStorage`.
+ */
+export const deserializeUTxOsFromStorage = (
+  serialized: Buffer,
+): Effect.Effect<
+  readonly UTxO[],
+  SDK.CborDeserializationError | SDK.CmlUnexpectedError
+> =>
+  Effect.gen(function* () {
+    const parsed = yield* Effect.try({
+      try: () => JSON.parse(serialized.toString("utf8")) as unknown,
+      catch: (e) =>
+        new SDK.CborDeserializationError({
+          message: `Failed to deserialize UTxO list payload`,
+          cause: e,
+        }),
+    });
+    if (
+      !Array.isArray(parsed) ||
+      parsed.some((entry) => typeof entry !== "string")
+    ) {
+      return yield* Effect.fail(
+        new SDK.CborDeserializationError({
+          message: `Invalid UTxO list payload`,
+          cause: parsed,
+        }),
+      );
+    }
+    return yield* Effect.forEach(
+      parsed,
+      (cborHex) =>
+        Effect.try({
+          try: () =>
+            coreToUtxo(
+              CML.TransactionUnspentOutput.from_cbor_bytes(
+                Buffer.from(cborHex, "hex"),
+              ),
+            ),
+          catch: (e) =>
+            new SDK.CmlUnexpectedError({
+              message: `Failed to deserialize UTxO from CBOR payload`,
+              cause: e,
+            }),
+        }),
+      { concurrency: "unbounded" },
+    );
+  });
